feat: delete FileProcessing.cs and implement it in Program.cs
